fixed type casting for accurate RMS calculation

+++ audio/dsp/modules/subbass_isolation.h
@@ -161,8 +161,8 @@ public:
             float sl = l_lp_.process(l_hp_.process(inL[i]));
             float sr = r_lp_.process(r_hp_.process(inR[i]));
 
-            sumL += sl * sl;
-            sumR += static_cast<double>(sr) * sr;
+            sumL += static_cast<double>(sl) * static_cast<double>(sl);
+            sumR += static_cast<double>(sr) * static_cast<double>(sr);
         }
 
         float rmsL = (n > 0) ? std::sqrt(sumL / n) : 0.0f;
